gstmppenc: fix broken FPS sanity check
The author seems unaware that FPS can be fractional, e.g. 7013/117
(59.94 Hz), so FPS_D isn't always = 1. The check was broken both
in only comparing the value of FPS_N and in not explicitly setting
FPS_D = 1 when trying to reset the FPS to 30

This code was causing excessive output bitrate by telling the
encoder to expect less than 1 FPS for such inputs, e.g. 30/117 for
a 7013/117 input (resulting in 233x higher bitrate than set)

Now only setting a 'default' FPS if there's no valid framerate
altogether. This may be the case if the framerate is not available
to the source element, but if wrong it will leave us encoding at
the wrong bitrate

Default FPS changed to 60 so if wrong we're more likely to encode
at lower than expected rather than higher bitrate